#335 に対応
