+++ src/Itinerary/Models/IDB01Itineraries/Contexts/ICT011ItineraryOne/ItineraryOne.tsx
@@ -15,31 +15,35 @@ export const ICT011ItineraryOneProvider = ({ children }: { children: ReactNode }
 		undefined,
 	);
 
+	const itineraryCRef = useMemo(
+		() =>
+			collection(db, Itineraries.modelName).withConverter(
+				FirestoreConverter<Itineraries, ItineraryOneInterface>(
+					Itineraries,
+					(data) => data,
+					(data) => data,
+				),
+			),
+		[],
+	);
+
 	useEffect(() => {
 		if (itineraryID) {
-			const unsubscribe = onSnapshot(
-				doc(collection(db, Itineraries.modelName), itineraryID).withConverter(
-					FirestoreConverter<Itineraries, ItineraryOneInterface>(
-						Itineraries,
-						(data) => data,
-						(data) => data,
-					),
-				),
-				(docSnap) => {
-					setItineraryDocSnap(docSnap);
-				},
-			);
+			const unsubscribe = onSnapshot(doc(itineraryCRef, itineraryID), (docSnap) => {
+				setItineraryDocSnap(docSnap);
+			});
 			return () => unsubscribe();
 		}
 		return undefined;
-	}, [itineraryID]);
+	}, [itineraryCRef, itineraryID]);
 
 	const value: ItineraryOneValInterface = useMemo(
 		() => ({
 			itineraryDocSnap,
 			setItineraryID,
+			itineraryCRef,
 		}),
-		[itineraryDocSnap, setItineraryID],
+		[itineraryDocSnap, setItineraryID, itineraryCRef],
 	);
 	return <ICT011ItineraryOne.Provider value={value}>{children}</ICT011ItineraryOne.Provider>;
 };
