+++ src/Itinerary/Pages/IPA001ItineraryEdit/ItineraryEdit.tsx
@@ -1,9 +1,8 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import { addDoc } from 'firebase/firestore';
-import React, { useCallback, useContext, useEffect } from 'react';
+import React, { useContext } from 'react';
 import { ScrollView, ActivityIndicator } from 'react-native';
 
-import { ItineraryOneInterface } from 'spelieve-common/lib/Interfaces/Itinerary/ICT011';
+import { IPA001ItineraryEditController } from './ItineraryEditController';
 
 import { BottomTabParamList } from '@/App';
 import { ICT011ItineraryOne } from '@/Itinerary/Models/IDB01Itineraries/Contexts/ICT011ItineraryOne';
@@ -15,38 +14,14 @@ export const IPA001ItineraryEdit = ({
 	route,
 	navigation,
 }: NativeStackScreenProps<BottomTabParamList, 'IPA001ItineraryEdit'>) => {
-	const { setItineraryID, itineraryDocSnap } = useContext(ICT011ItineraryOne);
-	const { isPlansLoading, plansDocSnapMap } = useContext(ICT031PlansMap);
-	const { planGroupsQSnap, planGroupsCRef } = useContext(ICT021PlanGroupsList);
+	const { itineraryDocSnap, itineraryCRef } = useContext(ICT011ItineraryOne);
+	const { isPlansLoading } = useContext(ICT031PlansMap);
+	const { planGroupsQSnap } = useContext(ICT021PlanGroupsList);
 
-	// eslint-disable-next-line @typescript-eslint/naming-convention
-	const { itineraryID, place_id, placeName } = route.params;
-
-	// TODO: 課題解消後Conroller に移動する
-
-	console.log({ itineraryDocSnap, isPlansLoading, planGroupsQSnap });
-	// TODO: https://github.com/Ayato-kosaka/spelieve/issues/335 Itinerary 新規作成
-	const createItinerary = useCallback(async () => {
-		console.log('createItinerary', itineraryDocSnap);
-		if (itineraryDocSnap) {
-			const itineray = await addDoc<ItineraryOneInterface>(itineraryDocSnap.ref.parent, {
-				title: '',
-				startDate: new Date(),
-				tags: [],
-				caption: '',
-				isUpdatable: true,
-				createdAt: new Date(),
-				updatedAt: new Date(),
-			});
-			navigation.setParams({ itineraryID: itineray.id });
-		}
-	}, [itineraryDocSnap, navigation]);
-
-	useEffect(() => {
-		if (itineraryID) {
-			setItineraryID(itineraryID);
-		}
-	}, [itineraryID, setItineraryID]);
+	const { createItinerary } = IPA001ItineraryEditController({
+		route,
+		navigation,
+	});
 
 	if (!itineraryDocSnap || isPlansLoading || !planGroupsQSnap) {
 		return <ActivityIndicator animating />;
